Move services to separate folder
