+++ src/app/core/services/favorite-cities/favorite-cities.service.spec.ts
