+++ src/app/dashboard/components/city-weather-card/city-weather-card.component.ts
@@ -1,6 +1,6 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { City } from 'src/app/core/models/City';
-import { FavoriteCitiesService } from 'src/app/core/favorite-cities/favorite-cities.service';
+import { FavoriteCitiesService } from 'src/app/core/services/favorite-cities/favorite-cities.service';
 
 @Component({
   selector: 'city-weather-card',
